Rewound to commit c43020374a3c0fb881bd0bfc60c2b87493a36175 + 0 edits

+++ index.js
@@ -26,8 +26,80 @@ client.on("ready", () => {
   );
 });
 
+client.on("message", message => {
+  if (message.author.bot) return;
+  const prefix = config.prefix;
+  const args = message.content
+    .slice(prefix.length)
+    .trim()
+    .split(/ +/g);
+  const command = args.shift().toLowerCase();
+  /*if (command === "pw") {
+    //Za to łatwo bana wyłapać :>
+    if (message.author.id == "688161946377257002") {
+      var msg = args.join(" ").toString();
+      message.guild.members.forEach(member => {
+        if (member.id != client.user.id && !member.user.bot) {
+          const embed = new Discord.RichEmbed()
+            .setColor("#FFFF00")
+            .setDescription(msg);
+          member.send(embed);
+        }
+      });
+      message.channel.send(
+        `Wysłano podaną wiadomość do ${
+          client.users.filter(user => !user.bot).size
+        } użytkowników`
+      ); //nie polecam używać na kanale głównym bo przyps, w dodatku kto by chciał coś wysyłać każdemu użytkownikowi na serwerze
+    } else {
+      message.channel.send("Nie masz permisji do użycia tej komendy!");
+    }
+  }*/
+
+  //unikałbym takiej składni w przyszłości, troche tu niepoukładanie
+  /*if (
+    message.mentions.users.get(client.user.id) ||
+    command === "help" ||
+    command === "pomoc"
+  ) {
+    message.channel.send(
+      "**LISTA KOMEND ADMINISTRACJI**\n```" +
+        config.prefix +
+        "PW - Wysyła do kazdego na serwerze wiadomość" +
+        config.prefix +
+        "powiedz - Wysyła wiadomość na danym kanale " +
+        config.prefix +
+        "status```\n**ADMINISTRACJA MOŻE TYLKO UŻYWAĆ!" +
+        client.users.find(user => user.id == "688161946377257002").tag +
+        "!**"
+    );
+  } else if (
+    message.content.startsWith("Cześć") ||
+    message.content.startsWith("Cześć!") ||
+    message.content.startsWith("Czesc")
+  ) {
+    message.channel.send("Cześć!");
+  } else if (
+    message.content.startsWith("Hej") ||
+    message.content.startsWith("Hej!")
+  ) {
+    message.channel.send("Hej!");
+  }*/
+
+  if (command === "powiedz") {
+    if (!message.content.startsWith(prefix)) return;
+    if (!message.author.id == "688161946377257002")
+      return message.channel.send("Nie masz permisji do użycia tej komendy!");
+    var wiadomosc = args.slice(0).join(" ");
+    if (!wiadomosc) return message.reply("Nie napisano żadnej wiadomości");
+    const embed = new Discord.RichEmbed()
+      .setColor("#ffff45")
+      .setDescription(wiadomosc);
+    message.channel.send(embed);
+  }
 
   if (command === "status") {
+    if (!message.content.startsWith(prefix)) return;
     if (!message.author.id == "688161946377257002")
       return message.channel.send("Nie masz permisji do użycia tej komendy!");
     message.delete();
@@ -38,6 +110,7 @@ client.on("ready", () => {
   }
 
   if (command == "ping") {
+    if (!message.content.startsWith(prefix)) return;
     var rola = message.member.roles.get("715511386809172038");
     var wiadomosc = args
       .slice(1)
@@ -52,21 +125,30 @@ client.on("ready", () => {
   }
 
   if (command == "tak") {
+    if (!message.content.startsWith(prefix)) return;
     message.member.addRole("715506096537600051");
   }
 
   if (command == "nie") {
+    if (!message.content.startsWith(prefix)) return;
     message.member.ban("Dostał wywalony ponieważ nie ma ukończone 13 lat ;(");
   }
 
   if (command == "rola-dodaj") {
-    var mention = message.mentions.roles.first();
-    function get(id) {
-      message.member.roles.get(id);
+    if (!message.content.startsWith(prefix)) return;
+    var roleMention = args
+      .slice(1)
+      .join(" ")
+      .toString();
+    var role = message.guild.roles.find(r => r.name == roleMention);
+    var memberMention = message.mentions.members.first();
+    if (!roleMention && !memberMention) return;
+    if (
+      !message.member.roles.get("715942031779692544") &&
+      !message.member.roles.get("715942039715446876")
+    ) {
+      return message.reply("Nie posiadasz permisji do użycia tej komendy");
     }
-    if (!mention) return;
-    if (!get("715942031779692544") && !get("715942039715446876")) 
-        return message.reply("Nie posiadasz permisji do użycia tej komendy"); //rola administratora // role moderator starszy moderator i v administrator te jeszcze
     /*var bannedRoles = [
       "715429731629006869",
       "715505723555184692",
@@ -86,21 +168,25 @@ client.on("ready", () => {
       "715546266033193031"
     ];
     if (bannedRoles.includes(mention.id)) return;*/
-    message.member.addRole(mention).then(member => {
-      message.channel.send(
-        `Użytkownik ${member.user.tag} otrzymał rolę ${mention.name}`
-      );
+    memberMention.addRole(role).then(member => {
+      const embed = new Discord.RichEmbed()
+        .setDescription(
+          `Użytkownik ${memberMention.user.tag} otrzymał rolę ${role.name}`
+        )
+        .setColor("#6bff00");
+      message.channel.send(embed);
     });
   }
 
   if (command == "propozycja") {
+    if (!message.content.startsWith(prefix)) return;
     var suggestion = args.join(" ").toString();
     const embed = new Discord.RichEmbed()
-      .setColor("#00ff00")
+      .setColor("#6bff00")
       .setAuthor(message.author.tag)
       .setDescription(suggestion);
     message.guild.channels
-      .get("715520608502546463")
+      .get("716573614979416115")
       .send(embed)
       .then(sentMessage => {
         sentMessage.react("👍");
@@ -109,18 +195,110 @@ client.on("ready", () => {
   }
 
   if (command == "skarga") {
+    if (!message.content.startsWith(prefix)) return;
     var skarga = args.join(" ").toString();
     const embed = new Discord.RichEmbed()
-      .setColor("#ff4d4d")
+      .setColor("#db8a8a")
       .setAuthor(message.author.tag)
       .setDescription(skarga);
     message.guild.channels
-      .get("715520630849667114")
+      .get("716573614979416115")
       .send(embed)
       .then(sentMessage => {
         sentMessage.react("👍");
         sentMessage.react("👎");
       });
+  }
+
+  if (command == "kick") {
+    if (!message.content.startsWith(prefix)) return;
+    if (
+      !message.member.roles.get(config.admID) &&
+      !message.member.roles.get(config.admID2)
+    )
+      return;
+    let member = message.mentions.members.first();
+    let powod = args
+      .slice(1)
+      .join(" ")
+      .toString();
+    console.log(`${member.tag} ${powod}`);
+    if (member == undefined || !powod) return;
+    member.kick(powod).then(kickedMember => {
+      console.log("kicked");
+      const embed = new Discord.RichEmbed()
+        .setDescription(
+          `Użytkownik ${member.tag} został wyrzucony przez ${message.author.tag} za: ${powod}`
+        )
+        .setColor()
+        .setTitle("BAN");
+      message.channel.send(embed);
+    });
+  }
+
+  if (command == "ban") {
+    if (!message.content.startsWith(prefix)) return;
+    if (
+      !message.member.roles.get(config.admID) &&
+      !message.member.roles.get(config.admID2)
+    )
+      return;
+    let member = message.mentions.members.first();
+    let powod = args
+      .slice(1)
+      .join(" ")
+      .toString();
+    console.log(`${member.user.tag} ${powod}`);
+    if (member === undefined || !powod) return;
+    member.ban(powod).then(bannedMember => {
+      console.log("banned");
+      const embed = new Discord.RichEmbed()
+        .setDescription(
+          `Użytkownik ${member.tag} został zbanowany przez ${message.author.tag} za: ${powod}`
+        )
+        .setColor()
+        .setTitle("BAN");
+      message.channel.send(embed);
+    });
+  }
+
+  if (command == "mute") {
+    if (!message.content.startsWith(prefix)) return;
+    if (
+      !message.member.roles.get(config.admID) &&
+      !message.member.roles.get(config.admID2)
+    )
+      return console.log("nie działa");
+    let muteRole = message.guild.roles.get("715512265620848681");
+    let powod = args
+      .slice(2)
+      .join(" ")
+      .toString();
+    let member = message.mentions.members.first();
+    let cooldown = parseInt(args[1]);
+    console.log(
+      muteRole.name + " " + powod + " " + member.user.tag + " " + cooldown
+    );
+    if (member.roles.get(muteRole.id)) return;
+    member.addRole(muteRole.id).then(muted => {
+      console.log("muted");
+      const embed = new Discord.RichEmbed()
+        .setDescription(
+          `Użytkownik ${muted.user.tag} został wyciszony przez ${message.author.tag} `
+        )
+        .setColor("#fffff0");
+      message.channel.send(embed);
+      member.send(
+        embed
+          .setDescription(
+            `Zostałeś wyciszony przez ${message.author.tag} za \${powod}\ `
+          )
+          .setFooter(member.user.avatarURL)
+      );
+      setTimeout(function() {
+        muted.removeRole(muteRole);
+      }, cooldown * 1000);
+    });
   }
 });
 client.login(config.token);
